Move Children into its own module (+0 B)
